backend/requirements.txt
django-ninja==0.19.1
environs==9.5.0
django-cors-headers==3.13.0